requirements.txt
coveralls==1.2.0
flake8==3.5.0
GitPython==2.1.8
mock==2.0.0
pytest==3.3.2
pytest-cov==2.5.1
